Revert "integration test: prevent GetMetric from interrupting the test when metric not found (#716)"
This reverts commit e9b92572d75ba87f327d29aed9631e59dd70685a.

The reason for revert: this change may break tests, which expect that ServesMetrics.GetMetric() fails
when the given metric doesn't exist in the output.

It is better to add 'TryGetMetric() (float64, bool)' function, which would return '(0, false)'
when the given metric doesn't exist, so the caller could decide what to do next.