lib/logstorage: return phrases with whitespace in front of uppercase letters from the appendCommonCasePhrases()
It is common to search for phrases with whitespace in front of captial letters.
For example, VictoriaMetrics is frequently written as 'Victoria Metrics', 'victoria metrics', 'Victoria metrics', etc.

So it is great from the usability PoV to search for these phrases with the `contains_common_case(...)` and `equals_common_case(...)` filters.

This is a follow-up for the commit a8358c033d17cefbc8d799117ed162ce1340973e