lib/logstorage: revert b41458ad1bd642500913424203f47274014b49e4
Reason for revert: the implicit whitespace delimiter in front of capital letters
is too subtle to understand when using contains_common_case() and equals_common_case() filters.
It also may create too many non-practical permutations of the input phrase, which could slow down the search.

The user can manually add a phrase with the needed whitespace delimiters into the list of phrases
passed to contains_common_case() and equals_common_case(). For example:

    contains_common_case("VictoriaMetrics", "Victoria Metrics")