Revert "deployment: upgrade libcrypto3 and libssl3 to 3.5.4-r0 (#736)"
This reverts commit 07da56ab0018f7145d98095ee406f0b2ff3a2e90.

Reason for revert: this change doesn't improve the security of Docker images for VictoriaLogs at all:
VictoriaLogs components are built as statically linked executables, so they do not depend
on external libraries. The "fix" updates external libraries inside the Alpine image to some pre-defined "secure"
versions with the following command at the Dockerfile, which builds base image for VictoriaLogs components:

    RUN apk add --no-cache 'libcrypto3=3.5.4-r0' 'libssl3=3.5.4-r0'

This complicates the stage for building the base Docker image, making it slower and more fragile.

As for the "security" scanners, which claim that this is big security breach, let's wait until this "breach"
is fixed upstream at Alpine, so later we could update our base image to the new Alpine release containing the "fix".

Updates https://github.com/VictoriaMetrics/VictoriaLogs/pull/736